+++ je_auto_control/utils/virtualized/__init__.py
@@ -1,0 +1,4 @@
+"""Realize off-screen items in virtualized lists / grids (ItemContainer + VirtualizedItem)."""
+from je_auto_control.utils.virtualized.virtualized import realize_item
+
+__all__ = ["realize_item"]